chore: refactor to use inheritance
