+++ postgrest_py/async_client.py
@@ -1,14 +1,16 @@
-from typing import Dict, Optional, Union
+from __future__ import annotations
+from typing import Dict, Optional, Type
+from types import TracebackType
 
 from deprecation import deprecated
-from httpx import AsyncClient, BasicAuth, Response
+from httpx import AsyncClient, Response
 
 from postgrest_py.__version__ import __version__
+from postgrest_py.base_client import BaseClient
 from postgrest_py.constants import DEFAULT_POSTGREST_CLIENT_HEADERS
-from postgrest_py.request_builder import RequestBuilder
 
 
-class AsyncPostgrestClient:
+class AsyncPostgrestClient(BaseClient):
     """Asyncio compatible PostgREST client."""
 
     def __init__(
@@ -23,54 +25,16 @@ class AsyncPostgrestClient:
             "Accept-Profile": schema,
             "Content-Profile": schema,
         }
-        self.session = AsyncClient(base_url=base_url, headers=headers)
+        self.session: AsyncClient = AsyncClient(base_url=base_url, headers=headers)
 
-    async def __aenter__(self):
+    async def __aenter__(self) -> AsyncPostgrestClient:
         return self
 
-    async def __aexit__(self, exc_type, exc, tb) -> None:
+    async def __aexit__(self, exc_type: Optional[Type[BaseException]], exc: Optional[BaseException], tb: Optional[TracebackType]) -> None:
         await self.aclose()
 
     async def aclose(self) -> None:
         await self.session.aclose()
-
-    def auth(
-        self,
-        token: Optional[str] = None,
-        *,
-        username: Union[str, bytes, None] = None,
-        password: Union[str, bytes] = "",
-    ) -> "AsyncPostgrestClient":
-        """
-        Authenticate the client with either bearer token or basic authentication.
-
-        Raise `ValueError` if neither authentication scheme is provided.
-        Bearer token is preferred if both ones are provided.
-        """
-        if token:
-            self.session.headers["apiKey"] = token
-            self.session.headers["Authorization"] = f"Bearer {token}"
-        elif username:
-            self.session.auth = BasicAuth(username, password)
-        else:
-            raise ValueError(
-                "Neither bearer token or basic authentication scheme is provided"
-            )
-        return self
-
-    def schema(self, schema: str) -> "AsyncPostgrestClient":
-        """Switch to another schema."""
-        self.session.headers.update({"Accept-Profile": schema, "Content-Profile": schema})
-        return self
-
-    def from_(self, table: str) -> RequestBuilder:
-        """Perform a table operation."""
-        return RequestBuilder(self.session, f"/{table}")
-
-    @deprecated("0.2.0", "1.0.0", __version__, "Use PostgrestClient.from_() instead")
-    def from_table(self, table: str) -> RequestBuilder:
-        """Alias to Self.from_()."""
-        return self.from_(table)
 
     async def rpc(self, func: str, params: dict) -> Response:
         """Perform a stored procedure call."""
@@ -80,8 +44,8 @@ class AsyncPostgrestClient:
 
 
 class Client(AsyncPostgrestClient):
-    """Alias to PostgrestClient."""
+    """Alias to AsyncPostgrestClient."""
 
-    @deprecated("0.2.0", "1.0.0", __version__, "Use PostgrestClient instead")
+    @deprecated("0.2.0", "1.0.0", __version__, "Use AsyncPostgrestClient instead")
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
